exception

+++ hexlet-spring-blog/src/main/java/io/hexlet/spring/controller/PostController.java
@@ -1,8 +1,11 @@
 package io.hexlet.spring.controller;
 
+import io.hexlet.spring.exception.ResourceNotFoundException;
 import io.hexlet.spring.repository.PostRepository;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,17 +33,22 @@ public class PostController {
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public Post createPost(@RequestBody Post post) {
-        return postRepository.save(post);
+    public ResponseEntity<Post> createPost(@Valid @RequestBody Post post) {
+        Post savedPost = postRepository.save(post);
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedPost);
     }
 
     @GetMapping("/{id}")
     public Post getPost(@PathVariable Long id) {
-        return postRepository.findById(id).orElse(null);
+        return postRepository.findById(id)
+                             .orElseThrow(() -> new ResourceNotFoundException("Post with id " + id + " not found"));
     }
 
     @PutMapping("/{id}")
-    public Post updatePost(@PathVariable Long id, @RequestBody Post post) {
+    public Post updatePost(@PathVariable Long id, @Valid @RequestBody Post post) {
+        if (!postRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Post with id " + id + " not found");
+        }
         post.setId(id);
         return postRepository.save(post);
     }
@@ -48,6 +56,9 @@ public class PostController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deletePost(@PathVariable Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User with id " + id + " not found");
+        }
         postRepository.deleteById(id);
     }
 }
